requirements.txt
pyodbc>=4.0.39
pandas>=2.0
python-dotenv>=1.0
SQLAlchemy>=2.0
scikit-learn>=1.1
